requirements.txt
Bootstrap-Flask==1.8.0
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-Session==0.4.0
Flask-WTF==0.14.3
PyMySQL==1.0.2
python-dotenv==0.18.0
